fix term filtering
